Create an optional second bridge
Some scenarios require a second ovs bridge to be used by
ovn. Here we check the existance of a file in `/etc/ovnk/extra_bridge`
and if exists, the content is expected to be the name of the interface
to be converted to the second bridge.

On top of that, in this scenario we detect and revert the short term
solution implemented by creating a single bridge which is not connected
to the default interface.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>